Parse JSON temperature to double instead of String->int

+++ app/src/main/java/com/shockn745/moovin5/motivation/recyclerview/cards/CardWeather.java
@@ -50,7 +50,7 @@ public class CardWeather extends AbstractCard {
 
     public CardWeather(Activity activity, FetchWeatherTask.WeatherInfos weatherInfos) {
         super(activity);
-        mTempText = Integer.toString(weatherInfos.mTemperature) + " °";
+        mTempText = Double.toString(weatherInfos.mTemperature) + " °";
         mForecastText = weatherInfos.mForecast;
         mImageResId = getIconResourceForWeatherCondition(weatherInfos.mWeatherId);
